backend/simple_speech_processor.py
#!/usr/bin/env python3
"""
Simple Speech Processor - From Scratch Approach
Simplified implementation focusing on getting basic functionality working
"""

import os
import tempfile
import logging
import hashlib
from typing import Optional
import speech_recognition as sr
from gtts import gTTS

class SimpleSpeechProcessor:
    def __init__(self):
        self.recognizer = sr.Recognizer()
        self.mock_mode = True  # Enable mock mode by default for testing

    def set_mock_mode(self, enabled: bool):
        """Enable or disable mock mode"""
        self.mock_mode = enabled
        logging.info(f"🧪 Mock mode {'enabled' if enabled else 'disabled'}")

    async def process_audio(self, audio_data: bytes) -> Optional[str]:
        """Main audio processing function - simplified approach"""
        try:
            logging.info(f"🎵 SimpleSpeechProcessor: Processing {len(audio_data)} bytes")

            # Quick size check
            if len(audio_data) < 100:
                logging.warning("❌ Audio too small, skipping")
                return None

            # If mock mode is enabled, return mock response for testing
            if self.mock_mode:
                return self._generate_mock_response(audio_data)

            # Try basic speech recognition
            result = await self._try_basic_recognition(audio_data)
            if result:
                return result

            # If that fails, fall back to mock
            logging.info("🔄 Basic recognition failed, falling back to mock")
            return self._generate_mock_response(audio_data)

        except Exception as e:
            logging.error(f"❌ SimpleSpeechProcessor error: {e}")
            # Always return something for testing
            return self._generate_mock_response(audio_data)

    def _generate_mock_response(self, audio_data: bytes) -> str:
        """Generate consistent mock responses for testing"""
        try:
            # Use file size and content hash to generate consistent responses
            size = len(audio_data)
            content_hash = hashlib.md5(audio_data).hexdigest()[:8]

            # Different responses based on size
            if size < 300:
                responses = [
                    "Hello, can you hear me?",
                    "Testing microphone.",
                    "This is a test.",
                    "How are you today?"
                ]
            elif size < 1000:
                responses = [
                    "Hello, I need help with something.",
                    "Can you tell me about your services?",
                    "What are your business hours?",
                    "I have a question about your products."
                ]
            else:
                responses = [
                    "Hello, I would like to know more about your company and the services you provide.",
                    "Can you help me understand how your products work and what they cost?",
                    "I'm interested in learning about your business and how you can help me.",
                    "What kind of support do you offer to your customers?"
                ]

            # Use hash to consistently select the same response for the same audio
            response_index = int(content_hash, 16) % len(responses)
            selected_response = responses[response_index]

            logging.info(f"🧪 Mock response (size: {size}, hash: {content_hash}): '{selected_response}'")
            return selected_response

        except Exception as e:
            logging.error(f"❌ Mock generation failed: {e}")
            return "Hello, this is a test message."

    async def _try_basic_recognition(self, audio_data: bytes) -> Optional[str]:
        """Try basic speech recognition"""
        temp_file = None
        try:
            # Save audio to temporary file
            with tempfile.NamedTemporaryFile(delete=False, suffix='.webm') as f:
                f.write(audio_data)
                temp_file = f.name

            # Try to recognize directly
            with sr.AudioFile(temp_file) as source:
                audio = self.recognizer.record(source)
                text = self.recognizer.recognize_google(audio)
                logging.info(f"✅ Basic recognition success: '{text}'")
                return text

        except Exception as e:
            logging.warning(f"⚠️ Basic recognition failed: {e}")
            return None
        finally:
            if temp_file and os.path.exists(temp_file):
                try:
                    os.unlink(temp_file)
                except:
                    pass

    async def text_to_speech(self, text: str) -> Optional[str]:
        """Convert text to speech - simplified"""
        try:
            # Generate TTS
            tts = gTTS(text=text, lang='en')

            # Save to file
            filename = f"simple_response_{hash(text) % 1000000}.mp3"
            file_path = os.path.join("backend/data/audio", filename)

            # Ensure directory exists
            os.makedirs(os.path.dirname(file_path), exist_ok=True)

            tts.save(file_path)
            logging.info(f"✅ TTS generated: {file_path}")
            return file_path

        except Exception as e:
            logging.error(f"❌ TTS failed: {e}")
            return None
