backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
requests==2.31.0
python-speech-recognition==3.10.0
pydub==0.25.1
gTTS==2.4.0
elevenlabs==0.2.26
google-generativeai==0.3.2
asyncio-mqtt==0.13.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
